requirements.txt
nextcord
pycryptodome
pytz
pyyaml
shortuuid
unidecode
